+++ src/main/java/com/psycrow/uncraftingtable/client/UncraftingTableScreen.java
@@ -9,24 +9,20 @@ import net.minecraft.resources.Identifier;
 import net.minecraft.world.entity.player.Inventory;
 
 public class UncraftingTableScreen extends AbstractContainerScreen<UncraftingTableMenu> {
-    private static final Identifier CRAFTING_TABLE_TEXTURE =
-            Identifier.withDefaultNamespace("textures/gui/container/crafting_table.png");
-    private static final Identifier CRAFTING_ARROW_SPRITE =
-            Identifier.withDefaultNamespace("container/crafting_table/arrow");
-
-    private static final int ARROW_X = 61;
-    private static final int ARROW_Y = 18;
+    private static final Identifier HOPPER_TEXTURE =
+            Identifier.withDefaultNamespace("textures/gui/container/hopper.png");
 
     public UncraftingTableScreen(UncraftingTableMenu menu, Inventory playerInventory, Component title) {
-        super(menu, playerInventory, title, 176, 166);
-        this.titleLabelX = 29;
+        super(menu, playerInventory, title, 176, 133);
+        this.inventoryLabelY = this.imageHeight - 94;
     }
 
     @Override
     public void extractBackground(GuiGraphicsExtractor graphics, int mouseX, int mouseY, float partialTick) {
+        super.extractBackground(graphics, mouseX, mouseY, partialTick);
         graphics.blit(
                 RenderPipelines.GUI_TEXTURED,
-                CRAFTING_TABLE_TEXTURE,
+                HOPPER_TEXTURE,
                 this.leftPos,
                 this.topPos,
                 0,
@@ -35,14 +31,5 @@ public class UncraftingTableScreen extends AbstractContainerScreen<UncraftingTab
                 this.imageHeight,
                 256,
                 256);
-        graphics.blitSprite(
-                RenderPipelines.GUI_TEXTURED,
-                CRAFTING_ARROW_SPRITE,
-                this.leftPos + ARROW_X,
-                this.topPos + ARROW_Y,
-                24,
-                17);
-
-        // Cycle button disabled until recipe cycling is re-enabled
     }
 }